Rename: User, Role entity 파일 위치 이동
