+++ src/main/java/com/mamoori/mamooriback/entity/Role.java
@@ -1,4 +1,4 @@
-package com.mamoori.mamooriback.oauth;
+package com.mamoori.mamooriback.entity;
 
 
 import lombok.Getter;
