+++ src/main/java/com/mamoori/mamooriback/repository/UserRepository.java
@@ -1,7 +1,8 @@
-package com.mamoori.mamooriback.oauth;
+package com.mamoori.mamooriback.repository;
 
 import java.util.Optional;
 
+import com.mamoori.mamooriback.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User,Long> {
