+++ src/main/java/com/mamoori/mamooriback/service/UserService.java
@@ -2,7 +2,7 @@ package com.mamoori.mamooriback.service;
 
 import java.util.Optional;
 
-import com.mamoori.mamooriback.oauth.User;
+import com.mamoori.mamooriback.entity.User;
 
 public interface UserService {
 	Optional<User> findByEmail(String email);
