+++ src/main/java/com/mamoori/mamooriback/service/impl/PostServiceImpl.java
@@ -4,10 +4,10 @@ import com.mamoori.mamooriback.controller.request.PostRequest;
 import com.mamoori.mamooriback.dto.PostResponse;
 import com.mamoori.mamooriback.entity.Post;
 import com.mamoori.mamooriback.entity.PostCategory;
-import com.mamoori.mamooriback.oauth.User;
-import com.mamoori.mamooriback.oauth.UserRepository;
+import com.mamoori.mamooriback.entity.User;
 import com.mamoori.mamooriback.repository.PostCategoryRepository;
 import com.mamoori.mamooriback.repository.PostRepository;
+import com.mamoori.mamooriback.repository.UserRepository;
 import com.mamoori.mamooriback.service.PostService;
 import com.mamoori.mamooriback.util.StringUtil;
 import lombok.RequiredArgsConstructor;
